rename files
renamed files and structured modules. fixed errors in project analysis
and projectutils. need to test on data at home
