+++ fmri_core/make_file_structure.sh
