+++ pythonutils/ridge/ridge.py
