+++ pythonutils/ridge/utils.py
