+++ pythonutils/strfval.py
